updates routes and api names
