+++ src/main/controllers/ChallengeController.kt
@@ -45,18 +45,6 @@ class ChallengeController: DefaultController<Challenge>(), RestController<Challe
         return SOAResult(SOAResultType.SUCCESS, challengesResult.message, challengesResult.data!!.challengeToUnsharedTransactions.map { it -> it.challenge })
     }
 
-    fun findAllChallengeBalances(user: UserAccount, requestData: RequestData): SOAResult<ChallengeToUnsharedTransactionsList> {
-        validateApiKey(user, requestData)
-
-        val challengesResult = ChallengeHelper.getChallenges(user)
-
-        if (challengesResult.data == null) {
-            throw InternalError()
-        }
-
-        return SOAResult(SOAResultType.SUCCESS, challengesResult.message, challengesResult.data!!)
-    }
-
     @Throws(ForbiddenException::class)
     fun complete(user: UserAccount?, requestData: RequestData): SOAResult<TransactionList> {
         validateApiKey(user!!, requestData)
@@ -123,7 +111,7 @@ class ChallengeController: DefaultController<Challenge>(), RestController<Challe
         return finalResult
     }
 
-    fun getAllBalancesForChallenge(user: UserAccount?, requestData: RequestData): SOAResult<EmailToChallengeBalanceList> {
+    fun balances(user: UserAccount?, requestData: RequestData): SOAResult<EmailToChallengeBalanceList> {
         validateApiKey(user!!, requestData)
 
         val challengeId = requestData.body["challengeId"] as Int
