+++ src/test/integration/handlers/UserIntegrationTest.kt
@@ -30,7 +30,7 @@ class UserIntegrationTest : WordSpec() {
         transaction {
             map = TestHelper.buildRequest(
                 user,
-                "/user/hello",
+                "/user_test/hello",
                 "POST"
             )
         }
